과정 등록학생 수정 2

+++ src/components/EnrolList.js
@@ -47,13 +47,9 @@ const EnrolList = (props) => {
     // useEffect : 컴퍼넌트 생명주기에 따라 DOM 렌더링 처리
     // props 객체에 값이 존재할때 마다 detailsList에 렌더링해서 화면에 표시
     useEffect(() => {
-        const curItemKey = props.studDetails.key;
-        if (curItemKey){
-            items = [...items, props.studDetails];
-            props.setStudDetails({});
-        }
         // 삭제 기능 수행
-        if (props.action === 'delete') {
+        // eslint-disable-next-line no-restricted-globals
+        if (props.action === 'delete' && confirm('정말로 삭제하시겠습니까?')) {
             // 삭제 대상 아이템을 키로 가져옴
             const deleteItem = items.filter(
                 (item) => item.key === props.selectedItemKey
@@ -63,7 +59,26 @@ const EnrolList = (props) => {
             // 삭제한 학생에 대한 참가가능 인원수 복구
             props.restoreSeats(deleteItem.program);
         }
-    } , [props]);
+        // ??
+
+        // 등록하기와 수정하기를 구분하는 조건 추가
+        const curItemKey = props.studDetails.key;
+        if (curItemKey){
+            // 전달받은 키와 리스트에서 일치하는 항목의 index를 알아냄
+            const idx = items.findIndex( (item) => item.key === curItemKey );
+
+            if (idx > -1) { // 키와 일치하는 항목이 리스트에 존재한다면
+                // 수정하기로 간주하고 해당항목에 대해 수정 작업 수행
+                items = items.map((item) =>
+                    item.key === curItemKey ? props.studDetails : item);
+
+            } else {  // 키와 일치하는 항목이 리스트에 존재하지 않으면
+                      // 등록하기로 간주하고 해당항목은 새로운 항목으로 취급
+                items = [...items, props.studDetails];
+            }
+            props.setStudDetails({});
+        }
+    });
 
     return (
         <div className="enrolList">
